+++ Changelog.md
@@ -3,8 +3,12 @@
 1. used [peplin/pygatt](https://github.com/peplin/pygatt) as backend 
 2. made _PyAnova_ object thread-safe
 
-# pyanova-[0.1.2]
+# ~~pyanova-[0.1.2]~~ recalled
 
 1. removed default debug logger
 2. fixed up the demo
 3. added this changelog
+
+# pyanova-[0.1.3]
+
+1. bug fix: fixed initialization bug, removed logger provider